api-server/requirements.txt
peewee
Flask
flask-cors
pyyaml
chromadb
tensorflow
google-cloud
google-generativeai
pillow
